tests/wpt/tests/html/browsers/browsing-the-web/navigating-across-documents/003.html

<!doctype html>
<title>Navigation from unload whilst traversing history</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="003-1.html"></iframe>
<script>
var t = async_test();

var pages = [];
var iframe = document.getElementsByTagName("iframe")[0];


onmessage = t.step_func(function(e) {
  pages.push(e.data);
  if(pages.length == 3) {
    assert_array_equals(pages, ["003-1", "003-2", "003-1"]);
    t.done();
    iframe.parentNode.removeChild(iframe);
  }
});
</script>
